- made adjustments to text input menu to work with scripts.

+++ src/d_dehacked.cpp
@@ -2293,7 +2293,7 @@ static int PatchStrings (int dummy)
 			holdstring.StripRight();
 			if (holdstring.Len() > 0 && holdstring[holdstring.Len()-1] == '\\')
 			{
-				holdstring.Truncate((long)holdstring.Len()-1);
+				holdstring.Truncate(holdstring.Len()-1);
 				Line2 = igets ();
 			}
 			else
